refactor: move Cluster definition DTO to core as ClusterDefinition
Move the serialization DTO from definition-model to core/cluster/model
and rename to ClusterDefinition. It is only used by ClusterAdapter for
repository serialization and does not need to be shared externally.
